+++ pyout/tests/test_elements.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 import pytest
 
 from pyout.elements import adopt
-from pyout.elements import StyleError
+from pyout.elements import StyleValidationError
 from pyout.elements import validate
 
 
@@ -49,7 +49,7 @@ def test_adopt():
 
 def test_validate_error():
     pytest.importorskip("jsonschema")
-    with pytest.raises(StyleError):
+    with pytest.raises(StyleValidationError):
         validate("not ok")